dar: simplify and generalize backup scripts
Now we use a single backup script: /usr/local/sbin/backup

It receives a sequence number: 0 for full backup, n>0 for incremental,
and "today" for the appropriate sequence number for today.

The sequence number for today is: day of the year (-1) modulo the backup
history, in days, that we want to preserve. Main advantage with respect
to previous approach (in addition to code refactoring) is that we are no
longer tight to sub-week periodicity. Any number of day is now a valid
backup recycle period.